Add Tailwind CSS and update API test component.
The changes introduced this commit include the settings and installations necessary to add Tailwind CSS to the project. These were done in the package.json, postcss.config.js, and tailwind.config.js. Tailwind CSS was added to have a more utility-first CSS framework.

There was also heavy refactoring in the ApiTest.vue component to reportedly fetch from an API and display users and their respective weather conditions. A new functionality was added to get weather data based on user location.