Update tests, add weather API and improve front-end fetch calls
This commit brings several enhancements to our tests and API:

Removed a basic test case in ExampleTest.php that checked for a 200 response at the root URL. This was done as the root URL no longer returns a meaningful response.

Several new test cases were added to ExampleTest.php which specifically check for a 200 response from the /users endpoint and the /weather endpoint. Another test checks to make sure the /users endpoint is returning the correct number of users.

A new file GetWeatherController.php is added which handles requests to the new /weather endpoint. This endpoint returns weather data for a given latitude and longitude using the OpenWeatherMap API. The data is then cached for 30 minutes to avoid excessive API calls.

In the front-end, ApiTest.vue now makes fetch calls to the new /users and /weather endpoints instead of making direct calls to the OpenWeatherMap API. This change simplifies the front-end code by delegating the complexity of handling the weather API to the back-end.

In the back-end, api.php was updated to define the new /users and /weather routes.

Lastly, the version of guzzlehttp/guzzle was bumped up from ^7.2 to ^7.5 in composer.json.